fix: handle Int overflow in ImageResource deserialization (#174)
The Readeck API can return Long.MIN_VALUE (-9223372036854775808) for
image width/height fields. This caused a JsonDecodingException on
deserialization, silently stopping the bookmark sync.

- Changed ImageResource DTO width/height from Int? to Long?
- Added safe coercion to Int in the mapper (clamp to 0..MAX_VALUE)
- Added test verifying deserialization of overflow image dimensions

+++ app/src/main/java/de/readeckapp/io/rest/model/BookmarkDto.kt
@@ -61,6 +61,6 @@ data class Resource(
 @Serializable
 data class ImageResource(
     val src: String,
-    val width: Int? = null,
-    val height: Int? = null,
+    val width: Long? = null,
+    val height: Long? = null,
 )
